Add user profile photo

+++ server/src/users/service.py
@@ -1,8 +1,13 @@
+import io
 import uuid
 from pathlib import Path
 
+from PIL import Image
 from sqlalchemy.exc import IntegrityError, NoResultFound
 
+from src.config import settings
+from src.s3.exceptions import CantDeleteFileFromStorage, CantUploadFileToStorage
+from src.s3.utils import delete_files, upload_file
 from src.users.enums import UserOrder
 from src.users.exceptions import (
     CantSubscribeToUser,
@@ -107,13 +112,17 @@ class UserService:
             return UserGet.model_validate(user)
 
         except IntegrityError as exc:
-            raise UsernameOrEmailAlreadyExists(f"User with username {data.username} already exists") from exc
+            raise UsernameOrEmailAlreadyExists(
+                f"User with username {data.username} already exists"
+            ) from exc
 
     async def delete_user(
         self,
         user_id: uuid.UUID,
     ) -> None:
         """Delete user by id."""
+        if not await self._delete_all_files_from_storage(user_id=user_id):
+            raise CantDeleteFileFromStorage("Failed to delete file from S3")
 
         await self._repository.delete(user_id=user_id)
 
@@ -128,7 +137,9 @@ class UserService:
             raise CantSubscribeToUser("Can't subscribe to yourself")
 
         try:
-            await self._repository.subscribe(user_id=user_id, subscriber_id=subscriber_id)
+            await self._repository.subscribe(
+                user_id=user_id, subscriber_id=subscriber_id
+            )
         except NoResultFound as exc:
             raise UserNotFound(f"User with id {user_id} not found") from exc
 
@@ -143,7 +154,9 @@ class UserService:
             raise CantUnsubscribeFromUser("Can't unsubscribe from yourself")
 
         try:
-            await self._repository.unsubscribe(user_id=user_id, subscriber_id=subscriber_id)
+            await self._repository.unsubscribe(
+                user_id=user_id, subscriber_id=subscriber_id
+            )
 
         except NoResultFound as exc:
             raise UserNotFound(f"User with id {user_id} not found") from exc
@@ -174,7 +187,84 @@ class UserService:
                 username=user.username,
                 subscribers_count=user.subscribers_count,
                 is_admin=user.is_admin,
-                is_subscribed=(curr_user and (curr_user.user_id in [u.user_id for u in user.subscribers])),
+                is_subscribed=(
+                    curr_user
+                    and (curr_user.user_id in [u.user_id for u in user.subscribers])
+                ),
             )
             for user in users
         ]
+
+    async def _delete_all_files_from_storage(
+        self,
+        user_id: uuid.UUID,
+    ) -> bool:
+        """Delete all user files from storage."""
+
+        return await delete_files(
+            filenames=[
+                settings.file_prefixes.profile_photo_small + str(user_id),
+                settings.file_prefixes.profile_photo_medium + str(user_id),
+                settings.file_prefixes.profile_photo_large + str(user_id),
+            ],
+        )
+
+    async def update_profile_photo(
+        self,
+        user_id: uuid.UUID,
+        photo: bytes,
+    ) -> bool:
+        """Update user profile photo."""
+
+        img_small = Image.open(photo)
+        img_medium = Image.open(photo)
+        img_large = Image.open(photo)
+
+        img_small.thumbnail((80, 80))
+        img_medium.thumbnail((160, 160))
+        img_large.thumbnail((240, 240))
+
+        img_small_bytes = io.BytesIO()
+        img_medium_bytes = io.BytesIO()
+        img_large_bytes = io.BytesIO()
+
+        img_small.save(img_small_bytes, "PNG")
+        img_small_bytes = img_small_bytes.getvalue()
+
+        img_medium.save(img_medium_bytes, "PNG")
+        img_medium_bytes = img_medium_bytes.getvalue()
+
+        img_large.save(img_large_bytes, "PNG")
+        img_large_bytes = img_large_bytes.getvalue()
+
+        await self._delete_all_files_from_storage(user_id)
+
+        if not (
+            await upload_file(
+                file=img_small_bytes,
+                filename=settings.file_prefixes.profile_photo_small + str(user_id),
+            )
+            and await upload_file(
+                file=img_medium_bytes,
+                filename=settings.file_prefixes.profile_photo_medium + str(user_id),
+            )
+            and await upload_file(
+                file=img_large_bytes,
+                filename=settings.file_prefixes.profile_photo_large + str(user_id),
+            )
+        ):
+            await self._delete_all_files_from_storage(user_id)
+            raise CantUploadFileToStorage()
+
+        return True
+
+    async def delete_profile_photo(
+        self,
+        user_id: uuid.UUID,
+    ) -> bool:
+        """Delete user profile photo."""
+
+        if not await self._delete_all_files_from_storage(user_id):
+            raise CantDeleteFileFromStorage()
+
+        return True
